requirements.txt
alabaster==0.7.13
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
astroid==2.15.5
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.5
crispy-bootstrap5==0.7
cryptography==40.0.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
dj-database-url==2.0.0
Django==4.1.9
django-allauth==0.54.0
django-anymail==10.0
django-cors-headers==4.0.0
django-coverage-plugin==3.0.0
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-extensions==3.2.1
django-model-utils==4.3.1
django-redis==5.2.0
django-render-partial==0.4
django-stubs==4.2.0
django-stubs-ext==4.2.0
djangorestframework==3.14.0
djangorestframework-stubs==3.14.0
docutils==0.19
drf-spectacular==0.26.2
exceptiongroup==1.1.1
executing==1.2.0
factory-boy==3.2.1
Faker==18.9.0
filelock==3.12.0
flake8==6.0.0
flake8-isort==6.0.0
gunicorn==20.1.0
hiredis==2.2.2
identify==2.5.24
idna==3.4
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.13.2
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
lazy-object-proxy==1.9.0
livereload==2.6.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==1.2.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
partial==1.0
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.2
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
pytest-sugar==0.9.7
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
render==1.0.0
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.4
stack-data==0.6.2
termcolor==2.3.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.2
traitlets==5.9.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
types-requests==2.30.0.0
types-urllib3==1.26.25.13
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.23.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
whitenoise==6.4.0
wrapt==1.15.0
